+++ CHANGELOG.md
@@ -21,8 +21,9 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 #### :bug: Bug Fix
 
-* Fixed closing the dropdown when clicking on an element with stop propagation `bSelect`
 * Fixed an issue with missing methods `element` and `elements` in the Block prototype `bSelect`
+* Fixed closing the dropdown when clicking on an element with stop propagation `bSelect`
+* Fixed getting a component in `getComponent` when an additional selector is passed `components/friends/dom`
 
 #### :house: Internal